[DOC] Add page for Saros version compatibility
Adds a dedicated page to the user documentation explaining the Saros
version compatibility. Also explains the special case when using a Saros
version containing the 'qualifier' element (e.g. development builds).

Adds the new compatibility page to the user documentation sidebar. As
this is kind of fringe knowledge, I added as the second-to-last entry
(only followed by the FAQ). Adding this information to the FAQ instead
of a separate section seemed too inaccessible/hard to find to me.

Adds a section on development builds to the release page. Explicitly
mentions that such builds may be unstable.